IGAPP-354: adjusted api-client imports, use styledComponents for MapView
