+++ native/ios/Integreat.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -1,0 +1,16 @@
+{
+  "object": {
+    "pins": [
+      {
+        "package": "MapLibre GL Native",
+        "repositoryURL": "https://github.com/maplibre/maplibre-gl-native-distribution",
+        "state": {
+          "branch": null,
+          "revision": "dc28d9ad9e1b5c52cdeab26e2fe1781db4974e9a",
+          "version": "5.11.0"
+        }
+      }
+    ]
+  },
+  "version": 1
+}
